Src/requirements.txt
confluent-kafka==2.4.0
pika==1.3.2
jupyter==1.0.0
pyyaml==6.0.1
